# docs/swagger.yaml
basePath: /api
definitions:
  database.BillingModel:
    enum:
    - prepaid
    - postpaid
    type: string
    x-enum-varnames:
    - BillingModelPrepaid
    - BillingModelPostpaid
...
  database.EntitlementEvent:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      deployment_id:
        type: string
      event_type:
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      subscription_id:
        type: string
      tenant_id:
        type: string
    type: object
...
  database.ListSubscriptionFeaturesRow:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      enabled:
        type: boolean
      feature_code:
        type: string
      feature_id:
        type: string
      limits:
        items:
          type: integer
        type: array
      name:
        type: string
      subscription_id:
        type: string
    type: object
  database.Subscription:
    properties:
      auto_renew:
        $ref: '#/definitions/pgtype.Bool'
      billing_model:
        $ref: '#/definitions/database.BillingModel'
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      end_date:
        $ref: '#/definitions/pgtype.Date'
      id:
        type: string
      plan_code:
        type: string
      start_date:
        $ref: '#/definitions/pgtype.Date'
      status:
        $ref: '#/definitions/database.SubscriptionStatus'
      tenant_id:
        type: string
      trial_end:
        $ref: '#/definitions/pgtype.Date'
      updated_at:
        $ref: '#/definitions/pgtype.Timestamptz'
    type: object
  database.SubscriptionStatus:
    enum:
    - trial
    - active
    - grace
    - past_due
    - cancelled
    - expired
    type: string
    x-enum-varnames:
    - SubscriptionStatusTrial
    - SubscriptionStatusActive
    - SubscriptionStatusGrace
    - SubscriptionStatusPastDue
    - SubscriptionStatusCancelled
    - SubscriptionStatusExpired
...
  handlers.AddSubscriptionFeatureRequest:
    properties:
      enabled:
        example: true
        type: boolean
      feature_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      limits:
        additionalProperties: {}
        type: object
    type: object
...
  handlers.CreateSubscriptionRequest:
    properties:
      auto_renew:
        example: true
        type: boolean
      billing_model:
        example: prepaid
        type: string
      end_date:
        example: "2027-01-01"
        type: string
      plan_code:
        example: enterprise
        type: string
      start_date:
        example: "2026-01-01"
        type: string
      status:
        example: trial
        type: string
      trial_end:
        example: "2026-02-01"
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.UpdateSubscriptionStatusRequest:
    properties:
      status:
        example: active
        type: string
    type: object
...
  pgtype.Bool:
    properties:
      bool:
        type: boolean
      valid:
        type: boolean
    type: object
  pgtype.Date:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: Cloud License Platform — multi-tenant SaaS licensing and entitlement
    management.
  title: CLP API
  version: "1.0"
paths:
...
  /subscriptions/{id}:
    get:
      description: Returns a subscription by UUID.
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a subscription
      tags:
      - subscriptions
  /subscriptions/{id}/cancel:
    post:
      description: Sets the subscription status to cancelled.
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Cancel a subscription
      tags:
      - subscriptions
  /subscriptions/{id}/events:
    get:
      description: Returns all entitlement events for the given subscription, newest
        first.
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.EntitlementEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List events for a subscription
      tags:
      - subscriptions
  /subscriptions/{id}/features:
    get:
      description: Returns the features and their enabled/limit state for the given
        subscription.
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.ListSubscriptionFeaturesRow'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List features for a subscription
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Upserts a feature entitlement for the given subscription (idempotent).
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      - description: Feature entitlement
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.AddSubscriptionFeatureRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Add or update a feature on a subscription
      tags:
      - subscriptions
  /subscriptions/{id}/features/{featureId}:
    delete:
      description: Removes the feature entitlement for the given feature ID from the
        subscription.
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      - description: Feature UUID
        in: path
        name: featureId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Remove a feature from a subscription
      tags:
      - subscriptions
  /subscriptions/{id}/status:
    patch:
      consumes:
      - application/json
      description: Sets the subscription status (trial, active, grace, past_due, cancelled,
        expired).
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateSubscriptionStatusRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update subscription status
      tags:
      - subscriptions
...
  /tenants/{id}/subscriptions:
    get:
      description: Returns all subscriptions for the given tenant, newest first.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List subscriptions for a tenant
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a new subscription for the given tenant.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a subscription for a tenant
      tags:
      - subscriptions
  /tenants/{id}/subscriptions/active:
    get:
      description: Returns the most recent subscription in trial, active, or grace
        status.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get active subscription for a tenant
      tags:
      - subscriptions
...
swagger: "2.0"